requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
curl_cffi==0.6.2
python-dotenv==1.0.1

# data / ml
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2
scikit-learn==1.4.2

# nlp
transformers==4.43.3
tokenizers==0.15.2

# tip: torch is installed separately by setup.bat via CPU wheel index:
# python -m pip install --index-url https://download.pytorch.org/whl/cpu torch torchvision torchaudio
